Add Dockerfile, cryptography requirement to testsite
Add the ability to run `testsite` as a docker image for development and
testing. Dockerfile retains the use of a `$VIRTUAL_ENV` for consistency
with pre-existing non-docker environments. The `venv` module is used
instead of the `virtualenv` standalone program to avoid the dependency
overhead of installing additional Debian packages. `Testsite` is
initialized to serve all IPs in order to work with Docker's network
proxy.

Add the `cryptography` module to testsite/requirements.txt to take
advantage of the space-saving and dependency resolution improvements of
pre-built wheel packages over Debian packages. See
[https://www.djaodjin.com/blog/packaging-python-app.blog.html]()