Support new Python and Django versions in unified requirements.txt
Supports:
    - Python 3.7 / Django 3.2 (maintains compatibility)
    - Python 3.9 / Django 4.2
    - Python 3.12 / Django 5.1

Add pytz as an explicit requirement to support existing code in Python
>= 3.9.  Upstream libraries have replaced dependency on `pytz' with
the `zoneinfo' module since it's addition to in the Python
standard library in version 3.9; they no longer bring in `pytz' as an
implicit dependency.

`Django-debug-toolbar' and `djangorestframework' were upgraded to
improve support for newer Django versions, address CVEs, and dodge
breaking changes.